moved MCRGsonUTCDateAdapter to mycore-base
